{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
django-extensions = "^3.2.1"
django-environ = "^0.10.0"
django-redis = "^5.2.0"
whitenoise = "^6.4.0"
pytz = "^2022.7.1"
redis = "^4.5.1"
argon2-cffi = "^21.3.0"


[tool.poetry.group.dev.dependencies]
django-querycount = "^0.8.3"
flake8 = "^6.0.0"
pylint = "^2.17.0"
pylint-django = "^2.5.3"
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
django-coverage-plugin = "^3.0.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
pytest-xdist = "^3.2.0"
pytest-watch = "^4.2.0"
django-dynamic-fixture = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  .*migrations/.*.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''